Move dispatchQueryEndpoint into metabase/api/query-endpoints
